+++ app/src/main/java/co/edu/udea/compumovil/gr01_20171/proyectoescuela/Vista/PantallaProfesor.java
@@ -50,11 +50,14 @@ public class PantallaProfesor extends AppCompatActivity {
         datos = OperacionesBaseDeDatos.obtenerInstancia(getApplicationContext());
         totalGrupos= datos.obtenerGruposDB();
 
-
-
     }
 
-
+    /**
+     * Organiza el ArrayList con los grupos tipo String para mostrarlos en un Alert Dialog
+     * y que el ususario pueda seleccionar el grupo que desee.
+     * @param a: Arraylist con los grupos a mostrar.
+     * @return Alert Dialog con los grupos contenidos en el.
+     */
     public AlertDialog listarGrupos(ArrayList<String> a){
         AlertDialog.Builder builder = new AlertDialog.Builder(PantallaProfesor.this);
         final CharSequence[] items = new CharSequence[a.size()];
@@ -72,7 +75,12 @@ public class PantallaProfesor extends AppCompatActivity {
 
         return builder.create();
     }
-
+    /**
+     * Toma los grupos que se tiene registrados hastta el momento y los procesa para mosrarlos
+     * como una cadena completa de Strings de la siguiente forma "1-a"
+     * @param a Entra un Arraylist de tipo grupo con los grupos almacenados actualmente.
+     * @return Arraylist con los grupos del tipo String.
+     */
     public ArrayList<String> convertirGrupos (ArrayList<Grupo> a){
         ArrayList<String> gruposs = new ArrayList<>();
         Grupo aux;
@@ -83,11 +91,19 @@ public class PantallaProfesor extends AppCompatActivity {
         return gruposs;
     }
 
+    /**Acciona y pasa a la siguiente actividad para agregar los grupos
+     *
+     * @param v : Recibe una vista
+     */
     public void clickAgregarGrupo(View v){
         Intent ingresar = new Intent(PantallaProfesor.this,DialogAgregarGrupos.class);
         startActivity(ingresar);
     }
 
+    /**
+     * Acciona el botón y permite registrar cada uno de los estudiantes del grupo
+     * @param v: Recibe una vista
+     */
     public void clickSubirEstudiantes(View v){
         grupos= datos.obtenerGruposDB();
         gruposString=convertirGrupos(grupos);
@@ -95,6 +111,10 @@ public class PantallaProfesor extends AppCompatActivity {
         dialog.show();
     }
 
+    /**
+     * Acciona el botón y permite pasa la actividad para agregar las materias a dictar.
+     * @param view: Recibe una vista
+     */
     public void clickMaterias(View view)
     {
         Intent intent = new Intent(this, MateriaAdicion.class);
