+++ app/src/main/res/layout/activity_agregar_estudiantes.xml
@@ -53,8 +53,10 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:hint="Identificación"
-                android:inputType="textCapWords|textPersonName"
-                android:textSize="14sp" />
+
+                android:textSize="14sp"
+                android:numeric="integer"
+                android:inputType="number"/>
         </LinearLayout>
 
         <LinearLayout
