+++ app/src/main/res/layout/activity_pantalla_configuracion.xml
@@ -6,12 +6,14 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     tools:context="co.edu.udea.compumovil.gr01_20171.proyectoescuela.Vista.PantallaConfiguracion">
+
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_gravity="center_vertical"
         android:layout_weight="2"
         android:orientation="horizontal">
+
         <LinearLayout
             android:layout_width="1dp"
             android:layout_height="match_parent"
@@ -24,19 +26,23 @@
             android:layout_height="match_parent"
             android:layout_weight="2"
             android:orientation="vertical">
+
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
                 android:layout_weight="1"></LinearLayout>
+
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
                 android:layout_weight="1">
+
                 <Button
-                    android:id="@+id/pruebita"
+                    android:id="@+id/btn_ver_grupos"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_gravity="center_horizontal"
+                    android:onClick="clickVerGrupos"
                     android:text="Ver grupos" />
             </LinearLayout>
 
@@ -46,15 +52,18 @@
                 android:layout_weight="0.6">
 
             </LinearLayout>
+
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
                 android:layout_weight="1">
+
                 <Button
                     android:id="@+id/btn_ingreso_profe"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_gravity="center_horizontal"
+                    android:onClick="clickAgregarElementos"
                     android:text="Agregar elementos" />
             </LinearLayout>
 
@@ -65,6 +74,7 @@
                 android:layout_weight="0.6">
 
             </LinearLayout>
+
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
